Change acmod_size to uint32_t (backport change in tboot-1.6)

+++ xmhf/src/xmhf-core/xmhf-bootloader/txt_acmod.c
@@ -321,7 +321,7 @@ uint32_t get_sinit_capabilities(acm_hdr_t* hdr)
     return info_table->capabilities._raw;
 }
 
-static bool is_acmod(void *acmod_base, size_t acmod_size, uint8_t *type,
+static bool is_acmod(void *acmod_base, uint32_t acmod_size, uint8_t *type,
                      bool quiet)
 {
     acm_hdr_t *acm_hdr;
@@ -333,7 +333,7 @@ static bool is_acmod(void *acmod_base, size_t acmod_size, uint8_t *type,
     if ( acmod_size < sizeof(acm_hdr_t) ) {
         if ( !quiet )
             printf("\t ACM size is too small: acmod_size=%x,"
-                   " sizeof(acm_hdr)=%x\n", (uint32_t)acmod_size,
+                   " sizeof(acm_hdr)=%x\n", acmod_size,
                    (uint32_t)sizeof(acm_hdr) );
         return false;
     }
@@ -349,7 +349,7 @@ static bool is_acmod(void *acmod_base, size_t acmod_size, uint8_t *type,
     if ( acmod_size != acm_hdr->size * 4 ) {
         if ( !quiet )
             printf("\t ACM size is too small: acmod_size=%x,"
-                   " acm_hdr->size*4=%x\n", (uint32_t)acmod_size,
+                   " acm_hdr->size*4=%x\n", acmod_size,
                    acm_hdr->size*4);
         return false;
     }
@@ -396,7 +396,7 @@ static bool is_acmod(void *acmod_base, size_t acmod_size, uint8_t *type,
     return true;
 }
 
-bool is_sinit_acmod(void *acmod_base, size_t acmod_size, bool quiet)
+bool is_sinit_acmod(void *acmod_base, uint32_t acmod_size, bool quiet)
 {
     uint8_t type;
